requirements.txt
google-cloud-storage==2.10.0  # For working with GCP Storage
google-cloud-secret-manager==2.16.0  # For accessing Secrets Manager
google-cloud-bigquery==3.13.0  # For BigQuery access and operations
pandas==2.1.1  # Core library for data manipulation and analysis
numpy==1.24.3  # NumPy for array operations (required by Pandas)
apache-airflow-providers-slack==9.0.0  # For sending messages to Slack
apache-airflow-providers-sftp==4.10.3 # For SFTP operations 
paramiko==2.9.0  # Required by SFTP provider
apache-airflow-providers-ssh==2.1.0 # Required for SFTP operations
asyncssh>=2.12.0 # Required for SFTP operations
apache-airflow-providers-amazon==9.1.0 # Required for S3 operations